move common test fixtures to common project
