+++ api/src/test/java/org/example/age/api/crypto/SecureIdTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.io.IOException;
-import org.example.age.testing.JsonTesting;
+import org.example.age.common.testing.JsonTesting;
 import org.junit.jupiter.api.Test;
 
 public final class SecureIdTest {
