+++ app/src/test/java/org/example/age/app/AppVerificationTest.java
@@ -15,8 +15,8 @@ import org.example.age.api.client.SiteApi;
 import org.example.age.api.crypto.SecureId;
 import org.example.age.app.config.AvsAppConfig;
 import org.example.age.app.config.SiteAppConfig;
-import org.example.age.testing.JsonTesting;
-import org.example.age.testing.TestClient;
+import org.example.age.common.testing.JsonTesting;
+import org.example.age.common.testing.TestClient;
 import org.example.age.testing.containers.TestContainers;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
