+++ common/src/test/java/org/example/age/common/AsyncCallsTest.java
@@ -1,8 +1,8 @@
 package org.example.age.common;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.example.age.testing.WebStageTesting.await;
-import static org.example.age.testing.WebStageTesting.awaitErrorCode;
+import static org.example.age.common.testing.WebStageTesting.await;
+import static org.example.age.common.testing.WebStageTesting.awaitErrorCode;
 
 import java.io.IOException;
 import okhttp3.ResponseBody;
