+++ common/src/test/java/org/example/age/common/testing/TestClientTest.java
@@ -1,4 +1,4 @@
-package org.example.age.testing;
+package org.example.age.common.testing;
 
 import static org.assertj.core.api.Assertions.assertThat;
