+++ common/src/testFixtures/java/org/example/age/common/testing/JsonTesting.java
@@ -1,4 +1,4 @@
-package org.example.age.testing;
+package org.example.age.common.testing;
 
 import static org.assertj.core.api.Assertions.assertThat;
