+++ module/client/src/test/java/org/example/age/module/client/AvsClientsConfigTest.java
@@ -1,7 +1,7 @@
 package org.example.age.module.client;
 
+import org.example.age.common.testing.JsonTesting;
 import org.example.age.module.client.testing.TestConfig;
-import org.example.age.testing.JsonTesting;
 import org.junit.jupiter.api.Test;
 
 public final class AvsClientsConfigTest {
