+++ module/crypto-demo/build.gradle.kts
@@ -19,5 +19,6 @@ dependencies {
 
     testAnnotationProcessor(libs.dagger.compiler)
 
+    testImplementation(testFixtures(project(":common")))
     testImplementation(project(":testing"))
 }
