+++ module/crypto-demo/src/test/java/org/example/age/module/crypto/demo/keys/EccPrivateKeyTest.java
@@ -1,8 +1,8 @@
 package org.example.age.module.crypto.demo.keys;
 
 import java.io.IOException;
+import org.example.age.common.testing.JsonTesting;
 import org.example.age.module.crypto.demo.testing.ConfigKeyPair;
-import org.example.age.testing.JsonTesting;
 import org.junit.jupiter.api.Test;
 
 public final class EccPrivateKeyTest {
