+++ module/request-demo/build.gradle.kts
@@ -13,6 +13,7 @@ dependencies {
 
     testAnnotationProcessor(libs.dagger.compiler)
 
+    testImplementation(testFixtures(project(":common")))
     testImplementation(project(":testing"))
     testImplementation(libs.dropwizard.core)
     testImplementation(libs.dropwizard.testing)
