+++ module/request-demo/src/test/java/org/example/age/module/request/demo/DemoAccountIdTest.java
@@ -17,9 +17,9 @@ import jakarta.ws.rs.core.MediaType;
 import java.io.IOException;
 import okhttp3.Request;
 import okhttp3.Response;
+import org.example.age.common.testing.TestClient;
 import org.example.age.service.module.request.AccountIdContext;
 import org.example.age.service.module.request.RequestContextProvider;
-import org.example.age.testing.TestClient;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
