+++ module/store-dynamodb/build.gradle.kts
@@ -18,6 +18,7 @@ dependencies {
 
     testAnnotationProcessor(libs.dagger.compiler)
 
+    testImplementation(testFixtures(project(":common")))
     testImplementation(project(":testing"))
     testImplementation(project(":test-containers"))
 }
