+++ module/store-redis/build.gradle.kts
@@ -17,6 +17,7 @@ dependencies {
 
     testAnnotationProcessor(libs.dagger.compiler)
 
+    testImplementation(testFixtures(project(":common")))
     testImplementation(project(":testing"))
     testImplementation(project(":test-containers"))
 }
