+++ module/store-redis/src/test/java/org/example/age/module/store/redis/client/RedisConfigTest.java
@@ -1,8 +1,8 @@
 package org.example.age.module.store.redis.client;
 
 import java.io.IOException;
+import org.example.age.common.testing.JsonTesting;
 import org.example.age.module.store.redis.testing.TestConfig;
-import org.example.age.testing.JsonTesting;
 import org.junit.jupiter.api.Test;
 
 public final class RedisConfigTest {
