+++ service/module/build.gradle.kts
@@ -15,6 +15,7 @@ dependencies {
 
     testAnnotationProcessor(libs.dagger.compiler)
 
+    testImplementation(testFixtures(project(":common")))
     testImplementation(project(":testing"))
     testImplementation(libs.okhttp.okhttp)
     testImplementation(libs.dropwizard.core)
