+++ service/module/src/test/java/org/example/age/service/module/request/RequestContextTest.java
@@ -17,7 +17,7 @@ import jakarta.ws.rs.core.MediaType;
 import java.io.IOException;
 import okhttp3.Request;
 import okhttp3.Response;
-import org.example.age.testing.TestClient;
+import org.example.age.common.testing.TestClient;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
