+++ service/src/test/java/org/example/age/service/AvsServiceConfigTest.java
@@ -1,8 +1,8 @@
 package org.example.age.service;
 
 import java.io.IOException;
+import org.example.age.common.testing.JsonTesting;
 import org.example.age.service.testing.TestConfig;
-import org.example.age.testing.JsonTesting;
 import org.junit.jupiter.api.Test;
 
 public final class AvsServiceConfigTest {
