+++ service/src/test/java/org/example/age/service/SiteServiceTest.java
@@ -2,8 +2,8 @@ package org.example.age.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.within;
-import static org.example.age.testing.WebStageTesting.await;
-import static org.example.age.testing.WebStageTesting.awaitErrorCode;
+import static org.example.age.common.testing.WebStageTesting.await;
+import static org.example.age.common.testing.WebStageTesting.awaitErrorCode;
 
 import dagger.Binds;
 import dagger.Component;
