+++ service/src/test/java/org/example/age/service/testing/TestWrappedAvsService.java
@@ -5,7 +5,7 @@ import org.example.age.api.AuthMatchData;
 import org.example.age.api.AvsApi;
 import org.example.age.api.VerificationRequest;
 import org.example.age.api.crypto.SecureId;
-import org.example.age.testing.WebStageTesting;
+import org.example.age.common.testing.WebStageTesting;
 
 /** Wrapper that converts uncaught exceptions to a failed stage. */
 public record TestWrappedAvsService(AvsApi delegate) implements AvsApi {
