+++ testing/src/main/java/org/example/age/testing/TestEnvModule.java
@@ -7,6 +7,7 @@ import jakarta.inject.Named;
 import jakarta.inject.Singleton;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import org.example.age.common.testing.TestObjectMapper;
 
 /**
  * Dagger module that binds...
